added mobile listener to epub page
